+++ __quickie/install.py
@@ -1,19 +1,14 @@
-from quickie import arg, command, script
+from quickie import arg, script
 
 
 @script
 @arg("--editable", "-e", action="store_true", help="Install in editable mode.")
 @arg("--dev", action="store_true", help="Install development dependencies.")
-def local(editable=False, dev=False):
+def install(editable=False, dev=False):
     editable = "-e" if editable else ""
     dev = "[dev]" if dev else ""
     return f"""
-    pipx install {editable} .{dev} --force
+    uv pip install {editable} .{dev}
     rm -rf build
     rm -rf src/quickie.egg-info
     """
-
-
-@command
-def dev():
-    return ["pip", "install", ".[dev]"]